Revert "Merge pull request #61 from ernestio/feature/157"
This reverts commit 4a31cc52db5c0d9014a19f5abf55d596dc759ba2, reversing
changes made to be29aceac4aa3b27d478de229b6a074ee286e6e1.
