+++ command/command.go
@@ -5,12 +5,10 @@
 package command
 
 import (
-	"fmt"
 	"github.com/ernestio/ernest-cli/manager"
 	"github.com/ernestio/ernest-cli/model"
 	"github.com/fatih/color"
 	"github.com/urfave/cli"
-	"log"
 )
 
 // setup ...
@@ -18,33 +16,10 @@ func setup(c *cli.Context) (*manager.Manager, *model.Config) {
 	config := model.GetConfig()
 	if config == nil {
 		config = &model.Config{}
-		if c.Command.Name != "target" && c.Command.Name != "setup" {
+		if c.Command.Name != "target" {
 			color.Red("Environment not configured, please use target command")
 		}
 	}
 	m := manager.Manager{URL: config.URL}
 	return &m, config
-}
-
-// askForConfirmation uses Scanln to parse user input. A user must type in "yes" or "no" and
-// then press enter. It has fuzzy matching, so "y", "Y", "yes", "YES", and "Yes" all count as
-// confirmations. If the input is not recognized, it will ask again. The function does not return
-// until it gets a valid response from the user. Typically, you should use fmt to print out a question
-// before calling askForConfirmation. E.g. fmt.Println("WARNING: Are you sure? (yes/no)")
-func askForConfirmation() bool {
-	var response string
-	_, err := fmt.Scanln(&response)
-	if err != nil {
-		log.Fatal(err)
-	}
-	okayResponses := []string{"y", "Y", "yes", "Yes", "YES"}
-	nokayResponses := []string{"n", "N", "no", "No", "NO"}
-	if containsString(okayResponses, response) {
-		return true
-	} else if containsString(nokayResponses, response) {
-		return false
-	} else {
-		fmt.Println("Please type yes or no and then press enter:")
-		return askForConfirmation()
-	}
 }
