+++ internal/features/cli/step_definitions.go
@@ -176,10 +176,6 @@ func init() {
 		}
 	})
 
-	And(`^File "(.+?)" does not exist$`, func(filename string) {
-		_ = os.Remove(filename)
-	})
-
 }
 
 func ernest(cmdArgs ...string) {
